server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-pptx==0.6.23
pillow==10.2.0
pydantic==2.6.0
requests==2.31.0
svglib==1.5.1
reportlab==4.0.9

